+++ CHANGELOG.md
@@ -6,6 +6,28 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 Currently, this repo is in Prerelease. When it is released, this project will adhere to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 ========
 
+## 0.10.4
+
+### Changes
+
+-   Updates the `Link` component's `linkType` prop to `type`, implements `React.forwardRef` to use with other libraries, and updates the stories.
+
+## 0.10.3
+
+### Changes
+
+-   Updates the `Icon` component's `name` prop to allow for `LogoNames` enum types.
+
+### Adds
+
+-   Adds a "negative" NYPL logo to be used on non-white backgrounds.
+
+## 0.10.2
+
+### Changes
+
+-   Updates the formatting of svg file names to use underscores
+
 ## 0.10.1
 
 ### Changes